test(cli): refactor to a central helpers location
