+++ packages/artillery/test/cli/command-report.test.js
@@ -1,5 +1,5 @@
 const tap = require('tap');
-const { execute, generateTmpReportPath } = require('../cli/_helpers.js');
+const { execute, generateTmpReportPath } = require('../helpers');
 
 tap.test('If we report specifying output, no browser is opened', async (t) => {
   const outputFilePath = generateTmpReportPath(t.name, 'html');
