+++ packages/artillery/test/cli/errors-and-warnings.test.js
@@ -1,5 +1,5 @@
 const tap = require('tap');
-const { execute } = require('../cli/_helpers.js');
+const { execute } = require('../helpers');
 const execa = require('execa');
 
 tap.test('GH #215 regression', async (t) => {
