+++ packages/artillery/test/cli/run-smoke.test.js
@@ -1,5 +1,5 @@
 const tap = require('tap');
-const { execute } = require('../cli/_helpers.js');
+const { execute } = require('../helpers');
 
 tap.test(
   'Running with no arguments prints out usage information',
