+++ packages/artillery/test/cloud-e2e/fargate/bom.test.js
@@ -1,10 +1,7 @@
 const { test, before, beforeEach } = require('tap');
 const { $ } = require('zx');
 const fs = require('fs');
-const {
-  generateTmpReportPath,
-  getTestTags
-} = require('../../cli/_helpers.js');
+const { generateTmpReportPath, getTestTags } = require('../../helpers');
 
 const A9_PATH = process.env.A9_PATH || 'artillery';
