+++ packages/artillery/test/cloud-e2e/fargate/memory.test.js
@@ -1,6 +1,6 @@
 const { test, before, beforeEach } = require('tap');
 const { $ } = require('zx');
-const { generateTmpReportPath, getTestTags } = require('../../cli/_helpers.js');
+const { generateTmpReportPath, getTestTags } = require('../helpers');
 
 const A9_PATH = process.env.A9_PATH || 'artillery';
