+++ packages/artillery/test/cloud-e2e/fargate/misc.test.js
@@ -2,7 +2,7 @@ const { test, before, beforeEach } = require('tap');
 const { $ } = require('zx');
 const chalk = require('chalk');
 const fs = require('fs');
-const { generateTmpReportPath, getTestTags } = require('../../cli/_helpers.js');
+const { generateTmpReportPath, getTestTags } = require('../helpers');
 
 const A9_PATH = process.env.A9_PATH || 'artillery';
