+++ packages/artillery/test/cloud-e2e/lambda/lambda-smoke.test.js
@@ -1,7 +1,7 @@
 const tap = require('tap');
 const fs = require('fs');
 const { $ } = require('zx');
-const { getTestTags, generateTmpReportPath } = require('../../cli/_helpers.js');
+const { getTestTags, generateTmpReportPath } = require('../../helpers');
 
 const tags = getTestTags(['type:acceptance']);
